+++ examples/fe-fpm-cli/src/index.ts
@@ -9,7 +9,7 @@ import inquirer from 'inquirer';
 import { create as createStorage } from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import { create } from 'mem-fs-editor';
-import { join } from 'path';
+import { join } from 'node:path';
 import { promisify } from 'util';
 
 const sampleAppPath = join(__dirname, '../sample/fe-app');