+++ packages/adp-tooling/test/unit/preview/descriptor-change-handler.test.ts
@@ -4,9 +4,9 @@ jest.mock('crypto', () => ({
 
 import type { Logger } from '@sap-ux/logger';
 import type { Editor } from 'mem-fs-editor';
-import * as crypto from 'crypto';
-import * as path from 'path';
-import * as fs from 'fs';
+import * as crypto from 'node:crypto';
+import * as path from 'node:path';
+import * as fs from 'node:fs';
 import {
     addAnnotationFile,
     addXmlFragment,