+++ packages/app-config-writer/src/smartlinks-config/utils.ts
@@ -2,7 +2,7 @@ import type { AxiosBasicCredentials } from 'axios';
 import { cyan } from 'chalk';
 import { render } from 'ejs';
 import type { Editor } from 'mem-fs-editor';
-import { join } from 'path';
+import { join } from 'node:path';
 import type { ServiceProvider } from '@sap-ux/axios-extension';
 import { createForAbap, createForDestination } from '@sap-ux/axios-extension';
 import { isAppStudio } from '@sap-ux/btp-utils';