+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/odata-cli
 
+## 0.17.15
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/axios-extension@1.22.10
+    -   @sap-ux/btp-utils@1.1.4
+
 ## 0.17.14
 
 ### Patch Changes