+++ packages/ui5-info/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-info
 
+## 0.12.4
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+
 ## 0.12.3
 
 ### Patch Changes